Add health check endpoint and remove expose directive
The health check endpoint '/health' has been added to the nginx configuration for uptime monitoring. This endpoint will return a 200 status code with a 'UP' message in JSON format. Meanwhile, the 'EXPOSE 80' directive was removed from the Dockerfile as it is considered a good practice not to expose ports in the Dockerfile but in runtime with the use of '-p' during 'docker run' command.