Update NGINX and Dockerfile for development and production
The NGINX server now listens on port 80 instead of port 8080. Also, the Dockerfile has been significantly restructured for local development and production usage. Notably, for local development, the Angular CLI's development server is used instead of an NGINX server. Moreover, the application is built in the environment that Angular CLI's server provides for production. A custom NGINX configuration is then used to serve the application.