Add functionality to manage new users in admin
This commit introduces several changes to bolster the user management capabilities of the admin interface.

In `users.component.ts`:
+ Added more imports, including MatTooltip, MatDialog, and MemberNewComponent.
+ Replaced hard-coded roles array with a reference to a constant in a newly created constants file.
+ Added a method to open a new dialog for creating a new member.

In `users.service.ts`:
+ Added a new method for sending an invitation to a new member.

In `users.component.html`:
+ Overhauled the HTML to add a button for creating a new member.

In `es.json`:
+ Added new transloco scope for adding a new user.

A new component `member-new` was created with its associated TypeScript and HTML files for handling the creation of a new user.

In `new-or-edit.component.ts`:
+ Cleaned up some dev logs.

Lastly, a new constants file was created to store role definitions, and styles.scss was updated to fix some dialog overflow issues.